:wrench: added overall analytics calculation to web demo
