+++ app/src/main/java/com/shub39/grit/widgets/habit_overview_widget/HabitOverviewWidgetReceiver.kt
@@ -14,11 +14,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.widgets.habitstreakwidget
+package com.shub39.grit.widgets.habit_overview_widget
 
 import androidx.glance.appwidget.GlanceAppWidget
 import androidx.glance.appwidget.GlanceAppWidgetReceiver
 
-class HabitStreakWidgetReceiver : GlanceAppWidgetReceiver() {
-    override val glanceAppWidget: GlanceAppWidget = HabitStreakWidget()
+class HabitOverviewWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = HabitOverviewWidget()
 }
