+++ app/src/main/java/com/shub39/grit/widgets/habit_week_chart_widget/HabitWeekChartWidget.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.shub39.grit.widgets.habitweekchartwidget
+package com.shub39.grit.widgets.habit_week_chart_widget
 
 import android.content.Context
 import android.os.Build
